+++ src/coordinates/interfaces/IEquatorialSphericalCoordinates.ts
@@ -1,5 +1,0 @@
-export default interface IEquatorialSphericalCoordinates {
-    rightAscension: number,
-    declination: number,
-    radiusVector: number,
-}