+++ src/coordinates/interfaces/IRectangularCoordinates.ts
@@ -1,5 +1,0 @@
-export default interface IRectangularCoordinates {
-    x: number,
-    y: number,
-    z: number,
-}