+++ src/planets/Neptune.ts
@@ -3,12 +3,14 @@ import {calculateVSOP87, calculateVSOP87Angle} from '../utils/vsop87Calc';
 import {getAsyncCachedCalculation} from '../cache/calculationCache';
 import {observationCalc} from '../utils';
 import {DIAMETER_NEPTUNE} from '../constants/diameters';
-import IEclipticSphericalCoordinates from '../coordinates/interfaces/IEclipticSphericalCoordinates';
+import {EclipticSphericalCoordinates} from '../coordinates/coordinateTypes';
 import {normalizeAngle} from '../utils/angleCalc';
 import {getApparentMagnitudeNeptune} from '../utils/magnitudeCalc';
 
 export default class Neptune extends Planet {
-    public async getHeliocentricEclipticSphericalJ2000Coordinates(): Promise<IEclipticSphericalCoordinates> {
+    protected name = 'neptune';
+
+    public async getHeliocentricEclipticSphericalJ2000Coordinates(): Promise<EclipticSphericalCoordinates> {
         return await getAsyncCachedCalculation('neptune_heliocentric_spherical_j2000', this.t, async () => {
             const vsop87 = await import('./vsop87/vsop87NeptuneSphericalJ2000');
 
@@ -20,9 +22,11 @@ export default class Neptune extends Planet {
         });
     }
 
-    public async getHeliocentricEclipticSphericalDateCoordinates(): Promise<IEclipticSphericalCoordinates> {
+    public async getHeliocentricEclipticSphericalDateCoordinates(): Promise<EclipticSphericalCoordinates> {
         return await getAsyncCachedCalculation('neptune_heliocentric_spherical_date', this.t, async () => {
-            const vsop87 = await import('./vsop87/vsop87NeptuneSphericalDate');
+            const vsop87 = this.useVsop87Short
+                ? await import('./vsop87/vsop87NeptuneSphericalDateShort')
+                : await import('./vsop87/vsop87NeptuneSphericalDate');
 
             return {
                 lon: normalizeAngle(calculateVSOP87Angle(vsop87.VSOP87_X, this.t)),