+++ src/planets/Planet.ts
@@ -4,8 +4,8 @@ import {
     spherical2rectangular
 } from '../utils/coordinateCalc';
 import AstronomicalObject from '../astronomicalObject/AstronomicalObject';
-import IRectangularCoordinates from '../coordinates/interfaces/IRectangularCoordinates';
-import IEclipticSphericalCoordinates from '../coordinates/interfaces/IEclipticSphericalCoordinates';
+import {RectangularCoordinates} from '../coordinates/coordinateTypes';
+import {EclipticSphericalCoordinates} from '../coordinates/coordinateTypes';
 import IPlanet from './interfaces/IPlanet';
 import {observationCalc} from '../utils';
 import {createSun} from '../sun';
@@ -20,7 +20,7 @@ import {
 } from '../utils/apparentCoordinateCalc';
 import {createTimeOfInterest} from '../time';
 import {getRise, getSet, getTransit} from '../utils/riseSetTransitCalc';
-import ILocation from '../earth/interfaces/ILocation';
+import {Location} from '../earth/LocationTypes';
 import {STANDARD_ALTITUDE_PLANET_REFRACTION} from '../constants/standardAltitude';
 import Mercury from './Mercury';
 import Venus from './Venus';
@@ -34,56 +34,56 @@ export default abstract class Planet extends AstronomicalObject implements IPlan
     private sun: Sun;
     private earth: Earth;
 
-    constructor(toi?: TimeOfInterest) {
+    constructor(toi?: TimeOfInterest, protected useVsop87Short: boolean = false) {
         super(toi);
 
         this.sun = createSun(toi);
         this.earth = createEarth(toi);
     }
 
-    public async getHeliocentricEclipticRectangularJ2000Coordinates(): Promise<IRectangularCoordinates> {
-        const {lon, lat, radiusVector} = await this.getHeliocentricEclipticSphericalJ2000Coordinates();
+    public async getHeliocentricEclipticRectangularJ2000Coordinates(): Promise<RectangularCoordinates> {
+        const coords = await this.getHeliocentricEclipticSphericalJ2000Coordinates();
 
-        return spherical2rectangular(lon, lat, radiusVector);
+        return spherical2rectangular(coords);
     }
 
-    public async getHeliocentricEclipticRectangularDateCoordinates(): Promise<IRectangularCoordinates> {
-        const {lon, lat, radiusVector} = await this.getHeliocentricEclipticSphericalDateCoordinates();
+    public async getHeliocentricEclipticRectangularDateCoordinates(): Promise<RectangularCoordinates> {
+        const coords = await this.getHeliocentricEclipticSphericalDateCoordinates();
 
-        return spherical2rectangular(lon, lat, radiusVector);
+        return spherical2rectangular(coords);
     }
 
-    public abstract async getHeliocentricEclipticSphericalJ2000Coordinates(): Promise<IEclipticSphericalCoordinates>;
+    public abstract getHeliocentricEclipticSphericalJ2000Coordinates(): Promise<EclipticSphericalCoordinates>;
 
-    public abstract async getHeliocentricEclipticSphericalDateCoordinates(): Promise<IEclipticSphericalCoordinates>;
+    public abstract getHeliocentricEclipticSphericalDateCoordinates(): Promise<EclipticSphericalCoordinates>;
 
-    public async getGeocentricEclipticRectangularJ2000Coordinates(): Promise<IRectangularCoordinates> {
+    public async getGeocentricEclipticRectangularJ2000Coordinates(): Promise<RectangularCoordinates> {
         const coordsPlanet = await this.getHeliocentricEclipticRectangularJ2000Coordinates();
         const coordsEarth = await this.earth.getHeliocentricEclipticRectangularJ2000Coordinates();
 
         return rectangularHeliocentric2rectangularGeocentric(coordsPlanet, coordsEarth);
     }
 
-    public async getGeocentricEclipticRectangularDateCoordinates(): Promise<IRectangularCoordinates> {
+    public async getGeocentricEclipticRectangularDateCoordinates(): Promise<RectangularCoordinates> {
         const coordsPlanet = await this.getHeliocentricEclipticRectangularDateCoordinates();
         const coordsEarth = await this.earth.getHeliocentricEclipticRectangularDateCoordinates();
 
         return rectangularHeliocentric2rectangularGeocentric(coordsPlanet, coordsEarth);
     }
 
-    public async getGeocentricEclipticSphericalJ2000Coordinates(): Promise<IEclipticSphericalCoordinates> {
-        const {x, y, z} = await this.getGeocentricEclipticRectangularJ2000Coordinates();
+    public async getGeocentricEclipticSphericalJ2000Coordinates(): Promise<EclipticSphericalCoordinates> {
+        const coords = await this.getGeocentricEclipticRectangularJ2000Coordinates();
 
-        return rectangular2spherical(x, y, z);
+        return rectangular2spherical(coords);
     }
 
-    public async getGeocentricEclipticSphericalDateCoordinates(): Promise<IEclipticSphericalCoordinates> {
-        const {x, y, z} = await this.getGeocentricEclipticRectangularDateCoordinates();
+    public async getGeocentricEclipticSphericalDateCoordinates(): Promise<EclipticSphericalCoordinates> {
+        const coords = await this.getGeocentricEclipticRectangularDateCoordinates();
 
-        return rectangular2spherical(x, y, z);
+        return rectangular2spherical(coords);
     }
 
-    public async getApparentGeocentricEclipticSphericalCoordinates(): Promise<IEclipticSphericalCoordinates> {
+    public async getApparentGeocentricEclipticSphericalCoordinates(): Promise<EclipticSphericalCoordinates> {
         let coords = await this.getLightTimeCorrectedEclipticSphericalCoordinates();
 
         coords = correctEffectOfAberration(coords, this.T);
@@ -92,19 +92,19 @@ export default abstract class Planet extends AstronomicalObject implements IPlan
         return coords;
     }
 
-    public async getTransit(location: ILocation): Promise<TimeOfInterest> {
+    public async getTransit(location: Location): Promise<TimeOfInterest> {
         const jd = await getTransit(this.constructor, location, this.jd0);
 
         return createTimeOfInterest.fromJulianDay(jd);
     }
 
-    public async getRise(location: ILocation): Promise<TimeOfInterest> {
+    public async getRise(location: Location): Promise<TimeOfInterest> {
         const jd = await getRise(this.constructor, location, this.jd0, STANDARD_ALTITUDE_PLANET_REFRACTION);
 
         return createTimeOfInterest.fromJulianDay(jd);
     }
 
-    public async getSet(location: ILocation): Promise<TimeOfInterest> {
+    public async getSet(location: Location): Promise<TimeOfInterest> {
         const jd = await getSet(this.constructor, location, this.jd0, STANDARD_ALTITUDE_PLANET_REFRACTION);
 
         return createTimeOfInterest.fromJulianDay(jd);
@@ -143,7 +143,7 @@ export default abstract class Planet extends AstronomicalObject implements IPlan
         return observationCalc.isWaxing(chi);
     }
 
-    private async getLightTimeCorrectedEclipticSphericalCoordinates(): Promise<IEclipticSphericalCoordinates> {
+    private async getLightTimeCorrectedEclipticSphericalCoordinates(): Promise<EclipticSphericalCoordinates> {
         const {radiusVector} = await this.getGeocentricEclipticSphericalDateCoordinates();
 
         const jd = getLightTimeCorrectedJulianDay(this.jd, radiusVector);
@@ -162,8 +162,8 @@ export default abstract class Planet extends AstronomicalObject implements IPlan
         const helRecEarthCoords = await this.earth.getHeliocentricEclipticRectangularDateCoordinates();
         const helRecPlanetCoords = await planet.getHeliocentricEclipticRectangularDateCoordinates();
 
-        const {x, y, z} = rectangularHeliocentric2rectangularGeocentric(helRecPlanetCoords, helRecEarthCoords);
+        const coords = rectangularHeliocentric2rectangularGeocentric(helRecPlanetCoords, helRecEarthCoords);
 
-        return rectangular2spherical(x, y, z);
+        return rectangular2spherical(coords);
     }
 }