+++ src/planets/Saturn.ts
@@ -3,12 +3,14 @@ import {getAsyncCachedCalculation} from '../cache/calculationCache';
 import {calculateVSOP87, calculateVSOP87Angle} from '../utils/vsop87Calc';
 import {observationCalc} from '../utils';
 import {DIAMETER_SATURN} from '../constants/diameters';
-import IEclipticSphericalCoordinates from '../coordinates/interfaces/IEclipticSphericalCoordinates';
+import {EclipticSphericalCoordinates} from '../coordinates/coordinateTypes';
 import {normalizeAngle} from '../utils/angleCalc';
 import {getApparentMagnitudeSaturn} from '../utils/magnitudeCalc';
 
 export default class Saturn extends Planet {
-    public async getHeliocentricEclipticSphericalJ2000Coordinates(): Promise<IEclipticSphericalCoordinates> {
+    protected name = 'saturn';
+
+    public async getHeliocentricEclipticSphericalJ2000Coordinates(): Promise<EclipticSphericalCoordinates> {
         return await getAsyncCachedCalculation('saturn_heliocentric_spherical_j2000', this.t, async () => {
             const vsop87 = await import('./vsop87/vsop87SaturnSphericalJ2000');
 
@@ -20,9 +22,11 @@ export default class Saturn extends Planet {
         });
     }
 
-    public async getHeliocentricEclipticSphericalDateCoordinates(): Promise<IEclipticSphericalCoordinates> {
+    public async getHeliocentricEclipticSphericalDateCoordinates(): Promise<EclipticSphericalCoordinates> {
         return await getAsyncCachedCalculation('saturn_heliocentric_spherical_date', this.t, async () => {
-            const vsop87 = await import('./vsop87/vsop87SaturnSphericalDate');
+            const vsop87 = this.useVsop87Short
+                ? await import('./vsop87/vsop87SaturnSphericalDateShort')
+                : await import('./vsop87/vsop87SaturnSphericalDate');
 
             return {
                 lon: normalizeAngle(calculateVSOP87Angle(vsop87.VSOP87_X, this.t)),