+++ src/planets/Uranus.ts
@@ -3,12 +3,14 @@ import {getAsyncCachedCalculation} from '../cache/calculationCache';
 import {calculateVSOP87, calculateVSOP87Angle} from '../utils/vsop87Calc';
 import {observationCalc} from '../utils';
 import {DIAMETER_URANUS} from '../constants/diameters';
-import IEclipticSphericalCoordinates from '../coordinates/interfaces/IEclipticSphericalCoordinates';
+import {EclipticSphericalCoordinates} from '../coordinates/coordinateTypes';
 import {normalizeAngle} from '../utils/angleCalc';
 import {getApparentMagnitudeUranus} from '../utils/magnitudeCalc';
 
 export default class Uranus extends Planet {
-    public async getHeliocentricEclipticSphericalJ2000Coordinates(): Promise<IEclipticSphericalCoordinates> {
+    protected name = 'uranus';
+
+    public async getHeliocentricEclipticSphericalJ2000Coordinates(): Promise<EclipticSphericalCoordinates> {
         return await getAsyncCachedCalculation('uranus_heliocentric_spherical_j2000', this.t, async () => {
             const vsop87 = await import('./vsop87/vsop87UranusSphericalJ2000');
 
@@ -20,9 +22,11 @@ export default class Uranus extends Planet {
         });
     }
 
-    public async getHeliocentricEclipticSphericalDateCoordinates(): Promise<IEclipticSphericalCoordinates> {
+    public async getHeliocentricEclipticSphericalDateCoordinates(): Promise<EclipticSphericalCoordinates> {
         return await getAsyncCachedCalculation('uranus_heliocentric_spherical_date', this.t, async () => {
-            const vsop87 = await import('./vsop87/vsop87UranusSphericalDate');
+            const vsop87 = this.useVsop87Short
+                ? await import('./vsop87/vsop87UranusSphericalDateShort')
+                : await import('./vsop87/vsop87UranusSphericalDate');
 
             return {
                 lon: normalizeAngle(calculateVSOP87Angle(vsop87.VSOP87_X, this.t)),