+++ src/planets/planetTypes.ts
@@ -1,0 +1,9 @@
+import TimeOfInterest from '../time/TimeOfInterest';
+
+export enum Position {'north', 'south'}
+
+export type Conjunction = {
+    toi: TimeOfInterest,
+    position: Position,
+    angularDistance: number,
+}