+++ src/stars/index.ts
@@ -1,0 +1,3 @@
+import * as createStar from './createStar';
+
+export {createStar};