+++ src/utils/observationCalc.ts
@@ -1,9 +1,9 @@
-import IEquatorialSphericalCoordinates from '../coordinates/interfaces/IEquatorialSphericalCoordinates';
+import {EquatorialSphericalCoordinates} from '../coordinates/coordinateTypes';
 import {deg2rad, normalizeAngle, rad2deg} from './angleCalc';
 
 export function getElongation(
-    equCoordsObj: IEquatorialSphericalCoordinates,
-    equCoordsSun: IEquatorialSphericalCoordinates
+    equCoordsObj: EquatorialSphericalCoordinates,
+    equCoordsSun: EquatorialSphericalCoordinates
 ): number {
     const raObjRad = deg2rad(equCoordsObj.rightAscension);
     const dObjRad = deg2rad(equCoordsObj.declination);
@@ -20,8 +20,8 @@ export function getElongation(
 }
 
 export function getPhaseAngle(
-    equCoordsObj: IEquatorialSphericalCoordinates,
-    equCoordsSun: IEquatorialSphericalCoordinates
+    equCoordsObj: EquatorialSphericalCoordinates,
+    equCoordsSun: EquatorialSphericalCoordinates
 ): number {
     const distObj = equCoordsObj.radiusVector;
     const distSun = equCoordsSun.radiusVector;
@@ -43,8 +43,8 @@ export function getIlluminatedFraction(phaseAngle: number): number {
 }
 
 export function getPositionAngleOfBrightLimb(
-    equCoordsObj: IEquatorialSphericalCoordinates,
-    equCoordsSun: IEquatorialSphericalCoordinates
+    equCoordsObj: EquatorialSphericalCoordinates,
+    equCoordsSun: EquatorialSphericalCoordinates
 ): number {
     const raObjRad = deg2rad(equCoordsObj.rightAscension);
     const dObjRad = deg2rad(equCoordsObj.declination);
@@ -68,4 +68,21 @@ export function getAngularDiameter(distance: number, trueDiameter: number): numb
     const delta = 2 * Math.atan2(trueDiameter, 2 * distance);
 
     return rad2deg(delta);
+}
+
+export function getAngularSeparation(
+    coords1: EquatorialSphericalCoordinates,
+    coords2: EquatorialSphericalCoordinates,
+): number {
+    const raRad1 = deg2rad(coords1.rightAscension);
+    const dRad1 = deg2rad(coords1.declination);
+    const raRad2 = deg2rad(coords2.rightAscension);
+    const dRad2 = deg2rad(coords2.declination);
+
+    const dRad = Math.acos(
+        Math.sin(dRad1) * Math.sin(dRad2)
+        + Math.cos(dRad1) * Math.cos(dRad2) * Math.cos(raRad1 - raRad2)
+    );
+
+    return rad2deg(dRad);
 }